Creating settings panel and beginnings of oauth hook
If a user navigates to wp-login.php and there is no query string,
or if there is `memberful_auth` present in the query string then
the plugin will attempt to do oauth authentication.

This is to allow admins etc. who won't be customers in memberful
to login to the wordpress site.