Fixing some bugs with syncing user accounts
Now correctly accounts for the situation where a member exists in the
wordpress install but does not have a valid mapping. Using 2 queries as
MySQL does not support FULL OUTER joins. That said, the second query
will only run if the mapping doesn't exist, which should only happen in
a very small set of cases.