Add basic API error logging to plugin #82
WordPress doesn't seem to have a built in transport for error logs, and
it's often hard for us to pinpoint the problems that are causing users
trouble as WP can just fail silently.

This commit introduces a basic error log that can be used for recording
failures from the Memberful HTTP API. Hopefully we can expand it to log
other Memberful related errors, and have the plugin auto-report them to
Memberful.