Do not use wp_insert_user for updating users
WordPress 4.0 introduced a change[0] that subtly changes the way
wp_insert_user() behaves. Before, you could pass an array of data to the
function, and it would not try to insert any fields that were not
present in the array.

Since the change, fields such as $user_pass are assumed to be blank if
not provided. wp_update_user() works around this by fetching the user
data first, then merging in the supplied changes.

fixes #113

[0]: https://core.trac.wordpress.org/changeset/28454