Remove debug error logging
We had issues with this code in several sites. `update_option` call was
failing with the error message: `Serialization of 'Closure' is not allowed`.

In an ideal case I would just solve this error and keep error reporting
in place. But in reality I wasn't able to reproduce this error,
therefore I couldn't fix it. Also, debug logging was kind of useful, but
not so much, because usually I was able to get admin access to affected
sites and I was able to solve issues without the debug log.