Fix the Restrict Content Tool for WooCommerce
The Restrict Content Tool was disabled on WooCommerce product pages.
However this doesn't make much sense because it can be disabled via
admin dashboard for a given product (or not enabled at all for). I think
that this filter was added by mistake and in real we wanted to disable
WooCommerce filters in order to make our Restrict Content Tool work with
WooCommerce.