Fix Memberful endpoint URLs
Our plugin used to generate Memberful endpoint URLs such as:

http://example.com?memberful_endpoint=auth

But in fact, this is an invalid URL. The correct version is:

http://example.com/?memberful_endpoint=auth

The only reason why this worked for such a long time is that browsers
and all other tools which used such URLs were smart enough to convert
the former version to the latter one.

However, it doesn't mean that we should continue generating invalid
URLs and this commit fixes it.