Always show user mappings in the debug view
The debug view shows user mappings only if number of total users is
different than number of mapped users. However, if there is an issue in
the mapping table and the previously mentioned condition is met, then we
can't see the mapping records and debug the issue.

This commit fixes it by always showing all mapping records.