Recover more reliably from duplicate mapping errors
When two requests attempt to sync the same new user at the same time,
the second typically encounters a "duplicate_user_for_member" error
propagated from the primary key constraint on the mapping table.

The previous handling attempted to recover by loading the user
inserted by the first transaction, but within the second transaction
that generally didn't work; it's likely that the repeatable read
isolation level prevented it.

This change switches to a rollback-and-retry approach. There are still
conceivable ways for the retry to fail (esp. if the first transaction
is still incomplete), but they're not very likely, so this seems like
an acceptable approach for now.