Add price to the memberful_buy_subscription_link shortcode
This includes the "Link to checkout" TinyMCE dialog for creating
shortcodes. The price field is shown for all plans at the moment. It
looks that it is possible to show it only for "Choose what you pay"
plans, however this change looks order of magnitude more complex and it
might not be worth doing it.

https://3.basecamp.com/3293071/buckets/17429243/todolists/2740201364