Fix bulk restrict category select
The JS to dynamically display form elements based on other input values
was not switching based on the element's property value, but was using
the element's attribute value set on page creation.

This would mean that when a select field was used as the dynamic element
for "data-depends-on" it would always return the value it was originally
set to for comparison, and not whatever it was updated to.