Allow filtering of private feed description
This commit introduces the `'memberful_private_rss_description'` filter
so that users can use custom descriptions in their feeds.

We can't pass `bloginfo_rss("description")` to the filter as it
[**always**][1] `echo`es the site's description, defeating the purpose of
having a filter.

Instead, we manually get the site's description (in RSS format) and pass
it to the new filter.

[1]: https://developer.wordpress.org/reference/functions/bloginfo_rss/