Add support for SG Optimizer frontend optimization
SiteGround has its own WordPress plugin for site optimization. One of
the things it does is that it optimizes (minifies) the frontend code.
However, this feature breaks our embed code.

This commit fixes it by adding a semicolon to the embed code. This fix
has been successfully tested on a production site.