Use configurable member role on pass reset
The plugin blocks use of the Wordpress password reset feature for
members, as this is something handled by Memberful directly.

We previously hard coded the "subscriber" role as being the one to block
when a password reset attempt is made via Wordpress, however we allow
the role given to Memberful members to be configurable, this change
ensures only the set role have access to password reset blocked.

https://github.com/memberful/memberful-wp/issues/279