Add checks for product-based restrictions
WIP: Completing support for all types of post restrictions by
plan/product or the global "all members" and "all subscribers" options.

This is the "expand" step of the refactor, pulling out all the required
logic in a very procedural, inefficient but readable style.

The next step is refactoring to smaller functions and getting more
reusable "DRY" code.