Fix marketing content for the restrict content box
This commit includes two changes:

1. Do not use `memberful_marketing_content()` in `memberful_wp_metabox()`

`memberful_marketing_content()` should be used only for the frontend. It
is not suitable for use in `memberful_wp_metabox()` (e.g. in the admin
area) because its value depends on actual post restriction settings.

If a post currently doesn't have any restrictions, then we fallback to
default marketing content even if we saved marketing content for the
post before. This commit fixes it by using
`memberful_post_marketing_content()` which always returns post's saved
marketing content no matter current restriction settings. In fact, this
is what we alrady do in `memberful_wp_add_term_metabox()` for term
metaboxes.

2. Apply the `memberful_marketing_content` filter only in `memberful_marketing_content()`

I don't think that we should use this filter for metaboxes. Actually, we
currently do it in `main` because we use `memberful_marketing_content()`
for the metabox, but I don't think that this is a correct behaviour.