Add content access by term (category or tag) (#275)
* Add content access by term (category or tag)

Restricting access to posts automatically by category or tag. With
related marketing content set at the term level.

Post restrictions take precedence, if any terms with restrictions are
added to a restricted post the post-specific restrictions will be used
and term restrictions ignored.

* Allow general restrictions options on terms

This change allows categories and tags to use the more general
restriction features that limit to "All members" or "Members with an
active subscription"

* Check for valid nonce before saving term metadata

* Fix "any active plan" restriction with no user

The "any active plan" restriction was only restricting content for
members with no plans, it should also restrict content when no member is
present.

* Make new ACL term function names clearer

* Show "Protected" tag for term restricted posts

The "Protected by Memberful" message will be displayed in the posts
admin index when a post is restricted by one of its tags or categories.

* Move cache of disallowed IDs out of shared method

The memberful_wp_user_disallowed_ids_from_acl function is used for
checking both the post and term ACLs, so can't cache the returned IDs
here.

This change moves this static definition into the calling functions.

* Restrict by term acl refactor (#296)

* Start ACL refactor with plans permission check

This updated code builds a list of plans that are required to view this
post, merges that with and plans required to view terms attached to the
post, and compares that with the plans the user has.

It still needs work to support restricting by product (download/podcast)
and the override options of "All members" and "Members with an active
subscription".

* Cache calls to check post access

Cache the result of a post access check for specific user, returning
early on cache hit. Also extracting a couple of functions for lookup of
access of posts and terms for any registered user or any user with
active subscription.

* Use post meta copy of ACL when checking plan

Avoid extra looping over all ACL rules by checking plans set for
specific post.

* Fix issue restricting posts to any registered user

This bug was resulting in more posts being restricted than should have
been.

* Remove caching of restriction checks (#297)

* Extract base ACL class (#298)

* Add text to the bulk restrict tab to explain the difference between the new category restrict feature

* Fix "Is user logged in?" check

The issue with `isset()` is that it returns `true` all the time because
`$user` is set to 0 if user is not logged in.

* Fix marketing content for the restrict content box (#300)

This commit includes two changes:

1. Do not use `memberful_marketing_content()` in `memberful_wp_metabox()`

`memberful_marketing_content()` should be used only for the frontend. It
is not suitable for use in `memberful_wp_metabox()` (e.g. in the admin
area) because its value depends on actual post restriction settings.

If a post currently doesn't have any restrictions, then we fallback to
default marketing content even if we saved marketing content for the
post before. This commit fixes it by using
`memberful_post_marketing_content()` which always returns post's saved
marketing content no matter current restriction settings. In fact, this
is what we alrady do in `memberful_wp_add_term_metabox()` for term
metaboxes.

2. Apply the `memberful_marketing_content` filter only in `memberful_marketing_content()`

I don't think that we should use this filter for metaboxes. Actually, we
currently do it in `main` because we use `memberful_marketing_content()`
for the metabox, but I don't think that this is a correct behaviour.

* Return only first term restricting a post

We do not need to build an array of all terms restricting a post because
we always use only the first term from this array. Therefore we can save
some work and return as soon as we find a term restring a post.

Co-authored-by: Patrik Bóna <mrhead@users.noreply.github.com>
Co-authored-by: Dan Carson <dan@okdan.com>
Co-authored-by: Patrik Bóna <patrik.bona@mrhead.sk>