Set current user after sign in
We have a customer who wants to redirect signed in members based on
their tier. If someone is a free member, then they should be redirected
to the free member dashboard. If someone owns Plan A, then they should
be redirected to Plan A dashboard, etc.

Unfortunately, this isn't possible at the moment.

We have a filter (`memberful_wp_after_sign_in_url`) which allows you to
modify the after sign in URL. However, our functions like
`is_subscribed_to_memberful_plan()` don't work in this filter because
current user is not set.

This commit fixes it by setting the current user immediately after sign
in.