Do not pass `$_REQUEST` and `$_SERVER` to underlying methods
`$_REQUEST` and `$_SERVER` are super globals. It means that they are
automatically available in all scopes.

We pass them to `Memberful_Wp_Endpoint` methods. It is unnecessary
because these variables are already available everywhere, and we also
don't need them most of the time.

Also, this introduces an indirection which makes security audits a bit
harder. Instead of looking for `$_REQUEST` and `$_SERVER` variables, we
need to follow the code flow, which is more complicated and error-prone.

This commit changes it by removing the indirection and using `$_REQUEST`
and `$_SERVER` directly where needed.