Typecast memberful_use_global_snippet params to int when updating
When caching is enabled, it is possible for the
`memberful_use_global_snippets` option to reset to `true` / `1` on the
next action.

When customers update the teaser content feature in the global marketing
content tab, we call `update_option (
'memberful_use_global_snippets', (int)
isset($_POST['memberful_use_global_snippets']));`. If the checkbox is
unticked, we set the `option_value` of `memberful_use_global_snippets`
to an empty string in the database since `isset` (somehow) returns an
empty string if the `$_POST` parameter does not exist.

An issue happens when we call `add_action( 'admin_init', 'memberful_wp_register_options' )`.
Inside `memberful_wp_register_options`, we loop through the options defined in `memberful_wp_all_options`
and call `add_option` for each entry.

One of these options include `'memberful_use_global_snippets' => TRUE,`
which enables the option every single time a WP Admin action runs since
we're calling it via the `admin_init` hook.

This issue only happens when caching is enabled, which makes me believe
it's an issue related to empty strings and caching via `add_options`.

I have attached a video of the mechanism in action in the PR
description.