Check if user exists before returning it
The plugin has a table that maps Memberful members to WP users. However,
when a WP user is deleted (like from the WP dashboard), we don't delete
the mapping.

When a webhook (or sign-in attempt) comes in for a member that is already
mapped, we try to fetch the WP user, and if it doesn't exist, we create
a new one.

This works most of the time, except with `member.deleted` webhooks,
where we don't check if the WP User exists, and it eventually raises an
error.

This commit fixes this by checking if the WP user exists before returning
the mapping. If it doesn't, since the member was deleted, we don't need
to create it again.

https://3.basecamp.com/3293071/buckets/5610905/card_tables/cards/7870254896